+++ src/star-search/schemas/render-on-client.schema.ts
@@ -1,0 +1,6 @@
+import { z } from "zod";
+
+export const RenderLottoFactorParams = z.object({
+  repoName: z.string(),
+});
+export type RenderLottoFactorParams = z.infer<typeof RenderLottoFactorParams>;