+++ src/star-search/star-search.module.ts
@@ -3,15 +3,17 @@ import { HttpModule } from "@nestjs/axios";
 import { OpenAIWrappedModule } from "../openai-wrapped/openai-wrapped.module";
 import { TimescaleModule } from "../timescale/timescale.module";
 import { RepoModule } from "../repo/repo.module";
-import { StarSearchService } from "./star-search.service";
 import { StarSearchController } from "./star-search.controller";
 import { StarSearchToolsService } from "./star-search-tools.service";
-import { BingSearchToolsSearch } from "./bing-search-tools.service";
+import { BingSearchAgent } from "./agents/bing-search.agent";
+import { PullRequestAgent } from "./agents/pull-request.agent";
+import { IssuesAgent } from "./agents/issues.agent";
+import { ReleaseAgent } from "./agents/releases.agent";
 
 @Module({
   imports: [HttpModule, TimescaleModule, OpenAIWrappedModule, RepoModule],
-  providers: [BingSearchToolsSearch, StarSearchService, StarSearchToolsService],
-  exports: [BingSearchToolsSearch, StarSearchService, StarSearchToolsService],
+  providers: [BingSearchAgent, IssuesAgent, PullRequestAgent, ReleaseAgent, StarSearchToolsService],
+  exports: [BingSearchAgent, StarSearchToolsService],
   controllers: [StarSearchController],
 })
 export class StarSearchModule {}